requirements.txt
# Web Framework
Flask==3.0.0
gunicorn==21.2.0

# GPU-accelerated ML
torch==2.5.1
torchaudio==2.5.1
faster-whisper==1.0.3

# Video Processing
yt-dlp==2024.11.4
opencv-python-headless==4.10.0.84
ffmpeg-python==0.2.0

# OCR
google-cloud-vision==3.8.0

# Utilities
requests==2.32.3
python-dotenv==1.0.1
redis==5.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Linting
ruff 